+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Fix: Make all events available in the prompt
 * Add more notes per task
 * Allow missing metadata for tasks
+* If date is null, remove datetime from the task
 
 ### 1.3.0: 2024-11-12